Separate performance_regtest into alloc and scan regtests.
